Implements custom Error type.
Switches to returning library specific error instead of panicking on
constructors due to environment or user passed data. Library code should
not panic in case of these cases. It is acceptable only in very rare non
user-controlled cases.

If on the other side an internal source provides e.g. non-UTF8 strings
the library will panic. This allows the interface to be more concise.